Arreglo de tipados y dependencias para microservicio de cuentas

+++ account-service/src/main/java/com/arqui/accountservice/dto/request/DiscountRequestDTO.java
@@ -8,5 +8,5 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class DiscountRequestDTO {
-    private Integer amount;
+    private Double amount;
 }
