+++ payment-service/src/main/java/com/arqui/paymentservice/repository/PaymentRepository.java
@@ -3,5 +3,5 @@ package com.arqui.paymentservice.repository;
 import com.arqui.paymentservice.entity.PaymentMethod;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PaymentRepository extends JpaRepository<PaymentMethod, Integer> {
+public interface PaymentRepository extends JpaRepository<PaymentMethod, Long> {
 }
